Refactor volume classes to own package.
