+++ src/main/java/de/wasenweg/komix/volumes/VolumeRepository.java
@@ -1,4 +1,4 @@
-package de.wasenweg.komix.comics;
+package de.wasenweg.komix.volumes;
 
 import java.util.List;
