+++ src/main/java/de/wasenweg/komix/volumes/VolumeService.java
@@ -1,6 +1,7 @@
-package de.wasenweg.komix.comics;
+package de.wasenweg.komix.volumes;
 
 import com.mongodb.client.result.UpdateResult;
+import de.wasenweg.komix.comics.Comic;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
